requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
blinker==1.5
blis==0.7.9
breadability==0.1.20
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
confection==0.0.3
contourpy==1.0.5
cycler==0.11.0
cymem==2.0.7
Cython==0.29.28
datasets==2.6.1
decorator==5.1.1
dill==0.3.5.1
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.8.0
flatbuffers==23.1.21
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.10.0
gast==0.4.0
gensim==4.2.0
ginza==5.1.2
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
ja-ginza==5.1.2
Janome==0.4.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
keras==2.11.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==15.0.6.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.9
nltk==3.7
numpy==1.23.3
oauthlib==3.2.2
openai==0.24.0
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
pandas-stubs==1.5.1.221024
pathy==0.6.2
Pillow==9.2.0
plac==1.3.5
preshed==3.0.8
protobuf==3.19.6
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pydantic==1.10.2
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rich==12.6.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.2
semver==2.13.0
sentencepiece==0.1.97
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
streamlit==1.13.0
SudachiDict-core==20221021
SudachiPy==0.6.6
sumy==0.11.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
thinc==8.1.5
threadpoolctl==3.1.0
tinysegmenter==0.4
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.0
tornado==6.2
tqdm==4.64.1
transformers==4.24.0
typer==0.4.2
types-pytz==2022.5.0.0
typing_extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wasabi==0.10.1
watchdog==2.1.9
Werkzeug==2.2.2
wordcloud==1.8.2.2
wrapt==1.14.1
xxhash==3.1.0
yarl==1.8.1
zipp==3.9.0
